// work-app/src/types/cardTypes.ts
export interface ItemType {
  id: number;
  image: string;
  aircraft_type?: string;
  vehicle_type?: string;
  type?: string;
  year?: number;
  floorspace?: number;
  seats?: number;
  beds?: number;
  city?: string;
  price: number;
  manufacturer?: string;
  model?: string;
  registration_number?: string;
  serial_number?: string;
  description?: string;
  area?: number;
  address?: string;
  baths?: number;
  parking?: number;
  VIN?: string;
  construction?: string;
}